requirements.txt
Django==1.7
Pillow==2.6.1
PyJWT==0.3.2
django-activity-stream==0.5.1
django-cors-headers==1.0.0
django-haystack==2.3.1
django-jsonfield==0.9.13
djangorestframework==2.4.4
elasticsearch==1.3.0
oauthlib==0.7.2
python-openid==2.2.5
python-social-auth==0.2.2-dev
pytz==2014.10
requests==2.5.0
requests-oauthlib==0.4.2
six==1.8.0
tldextract==1.5.1
urllib3==1.10
wsgiref==0.1.2
